+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Fixed an issue that inline attachments are shown multiple times [#145](https://github.com/TNG/JGiven/issues/145)
 * HTML Report: allow arbitrary protocols in custom navigation links [#146](https://github.com/TNG/JGiven/issues/146)
+* HTML Report: fixed an issue with attachments that could lead to the generation of attachments with identical names, overriding attachments of other scenarios [#144]((https://github.com/TNG/JGiven/issues/144)
 
 # v0.9.2